Handle pts fields as longs, so they don't wrap after 35 minutes
Quite likely the underlying cause of #146

+++ app/src/main/java/ie/macinnes/tvheadend/player/reader/PlainStreamReader.java
@@ -48,7 +48,7 @@ abstract class PlainStreamReader implements StreamReader {
 
     @Override
     public final void consume(@NonNull final HtspMessage message) {
-        final long pts = message.getInteger("pts");
+        final long pts = message.getLong("pts");
         final byte[] payload = message.getByteArray("payload");
 
         final ParsableByteArray pba = new ParsableByteArray(payload);
